fix file directory (name & path)
