+++ page/left-side/index.vue
@@ -1,34 +1,36 @@
 <script setup lang="ts">
-  const emit = defineEmits<{
-    (event: "toggleLeftSide"): void;
-  }>();
+import { ref, computed } from "vue";
 
-  const props = defineProps<{
-    isVisible: Boolean;
-  }>();
+const emit = defineEmits<{
+  (event: "toggleLeftSide"): void;
+}>();
 
-  const isShowSide = computed(() => props.isVisible);
+const props = defineProps<{
+  isVisible: Boolean;
+}>();
 
-  const toggleLeftSide = () => {
-    emit("toggleLeftSide");
-  };
+const isShowSide = computed(() => props.isVisible);
 
-  const historyItems = ref([
-    {
-      title: "Raising children",
-      isQT: false,
-      verse: "Ephesians 6:4",
-      summary: "Do not provoke others to anger.",
-      tags: ["Delight", "Judgement"],
-    },
-    {
-      title: "Raising children",
-      isQT: true,
-      verse: "Ephesians 6:4",
-      summary: "Do not provoke others to anger.",
-      tags: ["Delight", "Judgement"],
-    },
-  ]);
+const toggleLeftSide = () => {
+  emit("toggleLeftSide");
+};
+
+const historyItems = ref([
+  {
+    title: "Raising children",
+    isQT: false,
+    verse: "Ephesians 6:4",
+    summary: "Do not provoke others to anger.",
+    tags: ["Delight", "Judgement"],
+  },
+  {
+    title: "Raising children",
+    isQT: true,
+    verse: "Ephesians 6:4",
+    summary: "Do not provoke others to anger.",
+    tags: ["Delight", "Judgement"],
+  },
+]);
 </script>
 
 <template>
@@ -36,7 +38,12 @@
     <div v-if="isShowSide" class="left-side-container">
       <div class="search-bar-container">
         <SearchBar />
-        <IconComponent path="ic_arrow_right" :width="24" :height="24" @click="toggleLeftSide" />
+        <IconComponent
+          path="ic_arrow_right"
+          :width="24"
+          :height="24"
+          @click="toggleLeftSide"
+        />
       </div>
       <div class="history-list-container">
         <LeftSideHistoryList title="Yesterday" :historyItems="historyItems" />
@@ -51,48 +58,52 @@
 </template>
 
 <style lang="scss" scoped>
-  .left-side-container {
+.left-side-container {
+  position: fixed;
+  top: 0;
+  left: 0;
+  background-color: $background;
+  width: 100vw;
+  height: 100vh;
+  box-shadow: 2px 0 10px rgba(0, 0, 0, 0.2);
+  z-index: 1000;
+
+  .search-bar-container {
     position: fixed;
     top: 0;
     left: 0;
-    background-color: $background;
-    width: 100vw;
-    height: 100vh;
-    box-shadow: 2px 0 10px rgba(0, 0, 0, 0.2);
     z-index: 1000;
-
-    .search-bar-container {
-      position: fixed;
-      top: 0;
-      left: 0;
-      z-index: 1000;
-      background-color: $background;
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      padding: 16px;
-      padding-left: 12px;
-      gap: 8px;
-      width: 100%;
-    }
-
-    .history-list-container {
-      margin-top: 80px;
-      height: 100%;
-      padding-bottom: 264px;
-      overflow-y: auto;
-    }
-
-    .background-radial-gradient {
-      position: fixed;
-      bottom: 0;
-      left: 0;
-      width: 100%;
-      height: 320px;
-      z-index: 99999;
-      pointer-events: none;
-      background: radial-gradient(circle at 48% 100%, rgba(255, 255, 255, 0.6) 0%, rgba(165, 75, 179, 0) 90%);
-      opacity: 0.2;
-    }
+    background-color: $background;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 16px;
+    padding-left: 12px;
+    gap: 8px;
+    width: 100%;
   }
+
+  .history-list-container {
+    margin-top: 80px;
+    height: 100%;
+    padding-bottom: 264px;
+    overflow-y: auto;
+  }
+
+  .background-radial-gradient {
+    position: fixed;
+    bottom: 0;
+    left: 0;
+    width: 100%;
+    height: 320px;
+    z-index: 99999;
+    pointer-events: none;
+    background: radial-gradient(
+      circle at 48% 100%,
+      rgba(255, 255, 255, 0.6) 0%,
+      rgba(165, 75, 179, 0) 90%
+    );
+    opacity: 0.2;
+  }
+}
 </style>
