+++ page/prayer/index.vue
@@ -42,7 +42,7 @@ const sortingModalOpen = () => {
       <div class="prayer-content">
         <strong>April 28, 2025</strong>
         <div class="content-subject">Silent Whispers of the Morning</div>
-        <NuxtLink to="/home/prayer/detail">
+        <NuxtLink to="/prayer/id">
           <img src="/icon/ic_chevron_right.svg" alt="chevron-right" />
         </NuxtLink>
       </div>
